+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - SourceExpander.Embedder: Add NotEmbeddingSourceAttribute
 - SourceExpander.Embedder.Analyzer: Warning on expanding embedded source
+- Localize diagnostics
 ### Changed
 - SourceExpander.Embedder: Format EmbbedingSourceClass
 - SourceExpander.Embedder: Ignore compilation error